backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pdfplumber
python-docx
google-genai
